Fix SliderMultiplier/SliderTickRate parsing inaccuracy

+++ osu.Game/Beatmaps/BeatmapDifficulty.cs
@@ -29,8 +29,8 @@ namespace osu.Game.Beatmaps
             set => approachRate = value;
         }
 
-        public float SliderMultiplier { get; set; } = 1;
-        public float SliderTickRate { get; set; } = 1;
+        public double SliderMultiplier { get; set; } = 1;
+        public double SliderTickRate { get; set; } = 1;
 
         /// <summary>
         /// Maps a difficulty value [0, 10] to a two-piece linear range of values.
